requirements.txt
streamlit
numpy
Pillow
tensorflow